requirements.txt
altgraph==0.17.4
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.3
ttkbootstrap==1.10.1
urllib3==2.2.1
